feat: implement texture mapping
Credit to DirectXTK, we can load .dds file as shader resource in runtime. We
create 6 common static sampler binding to root signature. We use an SRV heap
to store SRVs. We bind this SRV heap to current command list. During the rendering pass, we set offset in the heap and grab different SRVs in the root
signature.