feat!: implement billboard sprite with GS
We define billboard vertex as a structure contains a float3 center position
and a float2 size. In vertex shader, we do nothing but pass value to GS. In
geometry shader, we append 4 vertices per billboard into the triangle stream.
After rasterization of triStream, we do the same old shading in PS. At the
setting of pso, we enable alpha to converge technique, which acquires a
muti-sampling render target.