feat: implement Gaussian Blur with compute shader
We design a effect class to manage 2 resources and computing step in post
process. Two textures perform as compute shader's input and output. In one
dimention(horz or vert), we dispatch ceil of (width or height / 256) groups, and
number of anthor dimention totally. In each thread group, thread share a piece
of cache memory to reduce the sampling. In hlsl, we first copy 256 + 2 *
blurRadius(including neighbor block's tail or head datas to perform blur on
boundary). Then wait for the group memory synchronization. After that we can
read input datas from cache, which save a lot of time.