feat: implement water surface simulation with CS
We use 3 single-channelled textures to represent surface height. We need
previous two frames' solution to compute next solution, so we ping-pong 3
textures to do so. In water rendering, IA inputs a smooth mesh surface, and we
use a displacement map in vertex shader to change the altitude and normal. No
need to doubt that GPU's simulation is much quicker than CPU.