feat: implement Sobel filter
Sobel filter is a edge detection algorithm which uses a 3x3 kernel to compute
gradient magnitude of a pixel. We use compute shader to generate the Sobel
output image in luminance, then render a full-screen rectangle, multiplying
render result and Sobel result. Therefore, we resolve the MSAA result to a midway texture and let it be the post-effect input.